Add files via upload
Added pentagons and squares. tried to fix the text that wouldn't go away after a few seconds. The text still keeps getting drawn to the window every frame and is flickering. I have tried a lot of different methods to remove it after a few seconds. The code still needs to be cleaned up a bit more but it works. Also Idk if the square and pentagon are supposed to look like that or if we are supposed to do some other formula for each.